Add very lazy Daisy Seed board
This just connects the jack and pot connector to a board with the Daisy
Seed on it.  Those connector locations need to be re-thought, and the
LEDs are basically randomly placed.  This is not going into an
enclosure, it's just getting tested next to the other boards with IDC
cables randomly hanging between them.  I'm not proud.

Note that routing could be improved by picking different pins from the
Daisy Seed, but this intentionally tries to be pin-compatible with the
Hothouse pedal, at the cost of making the routing trivially uglier.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>