Add board for just potentiometers
Look, with the larger 125B enclosure, this is pretty trivial, and I'd
rather have the electromechanical side separated out from everything
else.

And while the larger enclosure means that there's more room, I still
detest the fact that the 9mm pots are 10mm high with unused space below
them.  On the old 1590A boards I would place SMD components under them,
but here there's no point to even that.

So despite having more room, I'm trying a "low profile" configuration
where there are NPTH holes for the standoffs, and the through-holes for
the legs are big enough that it lays flat on the board and is only 7mm
high as a result.

Why? Just because.

If all goes well, I won't be messing with this board very much -
although I will have to think about the connector placement more when I
put the boards together and try it out.

Right now this just puts a SWD pin header (10-pin 1.27mm IDC cable) in
the middle of bottom of the board, which kind of defeats the point of
the whole "low profile" thing.  I might end up with a board-to-board
connector eventually.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>