Add initial 'Jacks' board for guitar pedal
This is the re-done base board from the 1590A project, but organized
differently: with the 1590A you had to have the audio jacks side-mounted
and the resulting 'base board' was the size of the whole enclosure and
also dealt with the stomp switch.

In contrast, in a new bigger 125B enclosure it makes a bit more sense to
top-mount the jacks, and the base board thus becomes just the 'jacks'
board that only deals with power and audio jacks.

The jacks board still deals with the signal switching, but it is now a
'soft switch' controlled by a logic-level signal.  I did explore that
option in the 1590A project, but now it's an integral part of the
design.

This only contains this 'jacks' board with some very basic power supply
reverse polarity protection, the high-impedance input signal buffering,
and the MOSFET switching logic for the output.

Note that you can choose to ignore the MOSFET switching, and do the
pass-through entirely on the effect board (for example by just having
the digital effect be the NULL effect with the resulting slight delay).

But an actual "true bypass" mode is not supported by this design: it's
not only not very interesting from an analog circuit design standpoint
(which is why I'm doing this all), it also would require passing off the
raw signal handling entirely to the other boards.

But I do expect the JFET buffer and the MOSFET switching to be quite
good based on previous experiments - it looks like complete overkill for
a guitar pedal.  Famous last words: I've only built a single LS844-based
JFET buffer so far, and while that one showed excellent results, maybe I
happened to be lucky.

The effect is expected to be handled by the Electrosmith Daisy Seed
board:

    https://electro-smith.com/products/daisy-seed

and the result is at least initially designed to be compatible with the
Hothouse pedal kit by Cleveland Music Co:

    https://clevelandmusicco.com/hothouse-diy-digital-signal-processing-platform-kit/

so none of this is in any way novel or interesting except as my
continual journey to learn electronics.

As part of that, the implementation is quite different from the Hothouse
pedal.  Part of it is in the switching logic, which the Hothouse does
not do, instead always going through the Daisy Seed (with the resulting
5.2ms latency even with effects disabled).  And part of it is in the
analog circuits, where the Hothouse pedal uses a regular (sane) op-amp
based buffering, while this uses discrete components.

I actually would like to learn the mixed signal side too, so one
possible longer-term plan is to actually integrate the STM32 and the
PCM3060 from the Daisy Seed onto the main board itself.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>