Add minimal stomp switch board
This is almost embarrassing, but I don't like solder lugs, and while all
it does is have a LED and a compact stomp switch, it's actually trying
to be subtly useful: you can populate one side or the other to change
the "handedness" of the connector, which is useful for dual stomp
situations.

And the 12mm LED distance should work for both the 1590A and 125B
enclosures.  Of course, that would require a 1590A board that does soft
switching.  To make that 1590A an option, this uses a 1.5mm JST-ZH
connector.

But for now it's hopefully useful at least as a "let's try this all out
by just wiring things together" thing.  Maybe I'll integrate it some
other way eventually, but at least for now I'm just doing this in small
modular pieces.

This just happens to be a particularly trivial modular piece.

NOTE! Despite being particularly trivial, the actual footswitch needs to
be attached the right way around, and the orientation isn't obvious.
Yes, there is a keyway slot on the footswitch that shows the
orientation, but it's easy to get wrong.

And perhaps more importantly, while the footprint is for the E-Switch
FS5700 footswitch that has good documentation, other footswitches seem
to have the same PCB layout but apparently different actuation.

For example, it appears that the Tayda A-1884 has the same footprint,
but numbers the pins differently (1-3 twice, rather than 1-6), and that
particular (bad) datasheet says that 2-3 are normally on, while pressing
the switch will connect 1-2.  Of course, the datasheet I found doesn't
make it clear which way is pin 1, so...

On the E-Switch one, it looks to be the other way around.

So be careful out there. Measure twice, solder once.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>