Add 1590A form factor base board design
This is *not* the same as the old 1590A project.  Instead it's basically
the 'Jacks' board, but in a 1590A form factor.

There is one big difference, though: the jacks board expects the effect
board to do the on/off toggle logic, since that's how the Daisy Seed is
set up.  But this 1590A board has the old effect board model where the
effect board only does the effect.

In fact the connectors to the effect board are the same, so old effect
boards should just work.

As well as they ever did, that is...

To handle the stomp switch, this re-introduces the on/off toggle circuit
I was toying with in the 1590A repo but never used, with the intention
of using the shared stomp switch board.

... except I didn't think this all through when I did that board, and
right now that board is set to always make the switch and LED connect to
GND, and so has only three wires.  That's what the Daisy Seed does, but
it's not what this "resurrect the 1590A form factor" board wants at all.

So this has a completely made-up four-pin connector that is most
definitely not the final one, and is not one that matches the current
stomp switch board.  It's a place-holder for when I make my mind up
about what I'll do.

But while the physical connector is wrong, it does have the right model
for the stomp switch connection: just two pairs of wires for the switch
and the LED respectively.

You could just solder individual wires like some kind of animal.  But
that's not how we roll, is it?

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>