Add 1590B symbol and footprint
Despite the README.md mention, this wasn't the enclosure I was thinking
of: the one I actually used for the jack board design was the slightly
bigger 125B enclosure (aka "Hammond 1590N1" - the naming really isn't
wonderful, which also explains my confusion).

However, there's a really nice 3D model STEP file for the 1590B that I'm
about to pull in by Aleksander Sadowski, so let's add the symbol and
footprint.

I've put the potentiometer and stomp box holes kind of randomly in here,
and I think that I might have to shrink the Jack board too to take the
smaller 1590B size into account.  But it would be good to aim for parts
that can fit in _either_ the 125B or the 1590B form factor.

Put another way: this all will likely result in more random layout
noise.  But I'll have to modify the stomp switch board anyway, so I'll
chalk this up to "early days yet", in addition to the usual "Linus is
disorganized and not very competent".

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>