Add the world's ugliest 1/4" mono jack 3d model
Ok, I admit it.  I'm not a designer, and I never will be.  But I like
openscad as a way to generate simple models from a simple textual
description.  And no, openscad doesn't spit out STEP files, but FreeCAD
does take openscad files and can export the result as a step file.

"Why don't you just use FreeCAD natively?" I hear you ask.

I find mousing around annoying.  I'd much rather write a simple text
file that specifies a few 3D shapes by where they are, and then generate
the 3d models from that.

And yes, they'll be ugly.  But they'd be ugly if I tried to use FreeCAD
too.  Because there's a common theme here.  Me.

Now, the FreeCAD import of openscad files looks fine inside FreeCAD, but
when it is then exported as a STEP file, the result is not great.
Instead of the "hole" for the 6.35mm jack, it's exported as a plug.

I suspect that what is going on is that it's probably done as a
"transparent plug", and then the transparency is lost at some point and
it just ends up being a solid plug rather than a hole.  But don't quote
me on that.

There's probably some trick to fix it, but I can't find it in myself to
care too deeply.  It makes a rather ugly part unnecessarily uglier, but
not enormously so.  The 3d model still works as a "this is the layout".

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>