Work around freecad import / export problems
It turns out that the STEP file generation is subtly influenced by odd
openscad file parsing issues.  I can re-organize the part to make it all
JustWork(tm).

So this avoids the "hole in the jack turns into a plug" issue by just
organizing the openscad file a bit differently.

In another iteration of trying out variations of a theme, things looked
fine in both openscad and freecad, but the STEP file export had the hole
not just as a plug, but missing the rotation.

So this is all a bit fragile and I'm not sure what the exact pattern to
the problem is, but at least this version seems to look reasonable.

"It compiles? Quick - ship it!"

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>