Add ugly 3d model to DC barrel jack
Based on the Würth datasheet:

   https://www.we-online.com/components/products/datasheet/694106301002.pdf

and my stumbling attempts at generating a STEP file with openscad and
freecad.

This time I couldn't get rid of the "hole turns into a plug" problem
with the freecad STEP file exprt, so once again the model looks fine in
both openscad and in freecad, but the STEP export just looks wrong with
no hole for the power plug.

There's some subtle order of operations bug thing going on.  But I'll
use it as a way to distract from the ugliness of the model: everybody
will just look at the lack of hole for the plug, and overlook the harsh
edges.

Also: dazzle them with colors.  That's the trick.  Nobody will question
my mad design skillz when it's colorized and there are obvious freecad
export issues.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>