Fix up DC power jack footprint
When doing the 3d model for the Würth 2.1mm DC power jack, I took the
leg widths from the datasheet, which says that the center pin is
connected to a leg that is slightly wider than the other ones (2.8mm vs
2.4mm).  So the 3d footprint reflects that.

And when then looking at the 3d viewer, the resulting 3d model leg then
didn't fit in the hole of the footprint.

Now, in reality that was becasue the 3d model legs were artificially
rather thicker than reality (at 0.6mm, when reality it is more like just
0.25mm thick), and the pcb footprint actually matched the recommended
hole pattern in that same datasheet (3mm x 0.8mm).

But looking around, the regular KiCad footprint for this thing does make
the hole for pin 1 bigger.  So this just makes my footprint match that,
since that's what I actually used to use on the PCB anyway (ie I
switched back to my own footprint only now that I added a 3d model for
it).

When I measure the actual DC jacks I have, they don't actually have that
pin as wide as 2.8mm.  But I have a mix of random DC jacks, so the one I
happened to measure may not even be a Würth one.

Put another way: this doesn't really matter, but for "match with
previous footprint" let's just fix it up.

In the new 1590A board I had randomly decided to test a surface-mount
version of this thing, so that doesn't have the same issue (and doesn't
have a 3d model at all right now as a result).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>