Update 1590A board
I really wanted to use the stomp switch board, but I didn't find any
sane connectors that are convenient and fit.

So I just updated this board to have the stomp switch directly on the
board instead.  I'd love to re-use stomp switches after I have made a
(bad) design and tested it, but the pain just doesn't seem worth it.

That means that I could just have used a traditional latching 3PDT, but
instead this sticks to the electronic switching with a momentary switch.
Because the whole point is the circuit, not sanity.

Add another openscad 3d model for this stomp switch thing.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>