Add 9mm potentiometer 3d model
This also shows how the 'Low profile' footprint uses the same model,
just shifted down by 3mm.

I hit more STEP file export problems, but seem to have worked around
them by saving in the AP242 format.  Or maybe it was some other random
change - the freecad STEP file export of OpenScad imports really is a
bit hit-and-miss.

This time the export randomly added another leg at 0,0 that didn't
actually exist.  I would expect that lots of people use the STEP file
export functionality, so I have this suspicion that it's the openscad
import that leaves some odd stale information around that then confuses
the export.

I was expecting to not hit any problems this time simply because there
was no 'difference()' operator, and that seemed to be what caused issues
before.  But clearly it's something else going on.

Whatever. The model looks fine now. Knock wood.

(And by "looks fine now", I obviously mean "it's as pretty as you could
ever expect my models to be")

I'm doing silly 3d models just to distract from the fact that I'm
frustrated with just physical layout issues.  I really don't love how
boring the 125B box is, and the 1590B box is actually smaller in the Z
dimension than a 1590A is.

The "Low Profile" use of the pot helps with the box height issue, and I
can easily fit things in a 1590B enclosure using the same model of
stacked PCBs that I used in the original 1590A setup.

But sane connectors are eluding me unless I just use pinheaders again,
and I was really hoping for a modular setup.  Some frankenstein thing
with 90-degree pin headers / sockets connecting boards side-to-side? In
addition to the vertical stacking? The mind boggles.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>