Add footprint for Neutrik NRJ6HM-1-PRE
I've been using that Tayda mono connector that is very compact, but I'm
playing around with this Neutrik version as a more robust component.  It
also happens to be stereo, which is irrelevant to me for now, but maybe
that will change some day.

Yes, there's already a footprint for this in the standard kicad
footprint libraries, but it's centered on the tip pin.

I find those kinds of footprints really annoying when the placement of
the component is about the physical placement of the _connector_, not
the pin, and that's what you really care about (ie for enclosure drill
files).

So this just does that Neutrik connector and centers it on the barrel of
the connector, so that you can just specify the component location by
where you want the barrel to stick out.  Which I feel is much more
relevant to any placement than where the pins then happen to end up.

Is it just me that does this? I've done the same for pots too: make the
footprint center on the shaft of the pot, not some random pin, and now
you can rotate the part around without changing the physical location of
the thing that actually matters, or just edit the exact placement in the
properties window.

I often wish many other components would be done that way too: I find it
annoying when I rotate a LED around to switch the legs around, and then
suddenly the LED is in a different position - because the part was
centered around a random leg rather than the part that matters.

Maybe there's some trick that real designers use that I'm missing, which
doesn't involve a calculator or eyeballing it.  But my solution has been
to just re-do the footprints the way I think they should have been done.

Also add a simplistic 3d model for this thing.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>