Update the Tayda A-859 stomp switch footprint
Ok, I have one of these in hand now, so I could verify the polarity of
the switch.  So rather than have holes for either orientation, just do
the correct side.

I probably actually want to use this switch, since a full-sized switch
is actually what I want for the double-decker model that I used for the
1590A enclosure, and I think that's the model I will go for in a 1590B
too.

Also do a 3d model for it.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>