Update 1590B enclosure footprint
This updates the 1590B footprint details, and adds Tayda-esque drill
patterns and a 3d-printed drill template for that layout.

The new layout uses side-mounted jacks rather than top-mounted ones,
kind of like the 1590A, except the jacks are symmetric rather than
offset (the 1590A required offset jacks just because they wouldn't fit
otherwise).

One reason to use side-mounted jacks is that the Neutrik jack wouldn't
fit top-mounted.  But I kind of like how it also separates the power
from the signal on the board instead of having everything crammed up
together at the top.

But I'm still playing around.  The drill template is good for
visualizing things, and I was even able to drill a passable enclosure
manually with it.  I'm still thinking about physical layout so this may
not be final, but it doesn't feel wrong.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>