Move 3d models to footprint directory
I tried to deal with the kicad path handling with a 'TOPLEVEL'
environment variable, but it worked horribly badly and got out of sync
because kicad didn't really integrate very well with it.

So just live with the fact that you can only have per-project paths for
the symbol and footprint libraries, not for the 3D models.  kicad does
have the notion of "3D model paths", but they aren't per project.

It looks like the way to keep kicad happy and avoid too much confusion
is to just make the 3D models be part of the footprint library, and put
the STEP files in the same path.  I'll keep the original openscad models
in the 'Models' (and 'Enclosures') directories.

Let's see if this fixes my issues.

Also, this mixes in various other edits to the 1590B footprint in
particular, since I couldn't keep the regularly scheduled kicad noise
separate from the 'move things around' noise.  kicad likes to randomly
update the uuid's of various parts as something changes (or, as I
sometimes feel, even if nothing changes).

So this is not a particularly clean commit: this also moves the pot
locations up by 5mm in the 1590B enclosure because that's what I was
looking at when the potentiometer board stopped showing 3d models due to
path confusion.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>