Make the 1590B 3d printer template have a copy of the board too
This uses the models to populate a rough copy of the actual board
layout, so that you get both a drill guide for drilling out the aluminum
enclosure, and a test print to add to it.

It also adds a model of the Daisy Seed to get some idea of placement.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>