Add 3D model for Daisy Seed
.. and also fix up the STEP file for the PBS-24-102P momentary stomp
switch.

It turns out that the trick for making freecad do proper STEP exports of
OpenScad imports seems to be to not do "select all" - which appears to
end up selecting some stale left-over results from the import too.

Instead do an OpenScad import, but then use the FreeCAD "box selection"
to select the 3d model, and then export that selection as a STEP file.
That seems to avoid the random odd stale leftovers from the import.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>