Another try at cleaning up project path handling
The removed TOPLEVEL hackery also tried to deal with the problem of
finding the spice models for simulation, which had the same confusion
with project-specific files and no sane kicad model that I could find
for it.

This attempts to deal with it by just making a 'spice' symlink from the
"kicad project" directory to the "real project" directory, and dealing
with the spice model paths that way.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>