Don't DC-couple the LS844 JFET input
I didn't think this through: it looks superficially fine only because
the circuit doesn't make the _other_ side of the JFET gate connection
obvious.

Without the AC-coupling capacitor, the input gate of the LS844 now has a
relatively low impedance path to ground, going through the instrument.
So the self-biasing to ground with a high-value resistor doesn't work at
all.

To do DC coupling, I'd need an actual negative rail.  I technically knew
that, but I clearly hadn't really internalized it and so just overlooked
this obvious (in hindsight) fact entirely.

I do have the simulation infrastructure for this and when you look at
the results you see that the biasing is all wrong without the
DC-blocking capacitor in place.

But to see that you do need to look at it, and I hadn't bothered since
it all used to work fine and I didn't think about this.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>