Update potentiometer board
I think I have decided on 2x04 connector to the potentiometer board, at
least for now.  It's a bit more mechanically stable than a 1x08.

So redo the routing, do some via stitching, and put that thing in the
middle of the potentiometer board.  I think I'm almost ready for a 1590B
Daisy Seed board.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>