Add 1590B base board design
This integrates the stomp switch and the Daisy Seed with the jacks on
the base board, because connecting things up was a bigger pain than just
integrating everything.

The end result mostly matches the 3d model in the Template directory.

As a result of the integration, this also removes the separate Jacks,
Daisy Seed and Stomp boards: they are all parts of the board here.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>