1590A: Replace the compact footswitch with a regular sized one
The bigger one probably fits better, since we have to reach up past the
audio jacks anyway.

I suspect I will go back to the actual 3PDT true bypass model for this
board, though.  The electronic switching is more interesting, but let's
face it, the simpler circuit with a physical switching of the signal is
simply more practical.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>