Add BC847C / BC857C transistors to my parts library
I didn't have a plain PNP BJT for some reason, and the dual transistor
packages I do have were a bit inconvenient when I wanted to use one
transistor on the signal path and another on the power path.  So add
another transistor to the list of transistors in my library.

And since I had looked at a high-h_{FE} version of the BC846 NPN BJT
that I already have, I decided to just go for it and add both a NPN and
PNP version.

The BC847C is a higher-h_{FE} grade of the BC847 series, which in turn
is similar to the BC846 but with lower V_{CEO} (45V vs 65V - irrelevant
for my use).

The BC857C is then the PNP version of that.

Do I need the higher DC gain? No.  I had looked at it due to my use of
the BC846 in that analog rail filter trial using the common collector
series follower setup.  But they are pennies, so why not?

Also add in the spice model for the NMB2227A, which was missing.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>