Add 'plain' 1590A base
This is the one that uses a traditional 3PDT stomp switch with true
bypass and thus avoids the whole soft-switching MOSFET games.

But to not make it too boring, I'm trying a different input buffer,
which is a bastardized Sziklai pair with a JFET and a PNP BJT.  It
simulates quite well, which obviously means not a whole lot.

This also uses a high-gain BC847C signal transistor to push the whole
attempted low-pass filter on the power rail even more.  That, in turn,
it driven almost entirely by the crazy simulated 60Hz ripple at 100mV
PtP.

Honestly, if you have 100mV ripple on your guitar power supply, you have
bigger issues and should just get a different power supply, but it was
fun to simulate this down to a 10uV signal with that 100mV power supply
ripple.

Of course real noise is very different, and no longer at 60Hz.  Modern
power supplies have that 10mV noise at a few hundred kHz (or in the MHz
range) instead.  So this is all pointless except as a trial simulation.

That JFET/PNP Sziklai pair also may not actually work in reality, but
hey, it's interesting enough that I think I'll at least try it out to
see.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>