Update CK-6.35 audio jack symbol, model and footprint
Fix up the CK-6.35 footprint for the correct symbol

Ok, annoyingly, I've been using the wrong audio jack symbol for the mono
case.  It works fine, but it wasn't really right.  I need to go back and
fix the schematics, but here's the right S/T/TN footprint, at least.

This also adds a "no non-plated through holes" version of the footprint:
it uses the small plastic legs as actual legs instead of as positioning
nubs.

That makes the part take a bit more vertical space (or you have to just
snip the plastic extrusions off, like I used to do) but can make it
easier to fit other components on the other side when there is no NPTH.

Finally, related to that "actual legs" change, this updates the 3d model
to have the right height for those feet and does a new freecad STEP file
export, and then adjusts the Z offset for the model in the footprint
version without the board holes.

This is just prep-work for any tight boards.  I'm looking at you,
1590LB.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>