Switch pots around for sane behavior
My 1590B test board works fine, but I was confused about the behavior of
the pots.  It turns out that for about the fifteenth time, I had them
the wrong way around.

Obviously the real cause is my incompetence, but I think it's made worse
by how the default symbol is laid out, with pin 1 at the top.  I then
mindlessly hook it up with positive rail at the top and ground at the
bottom, and the end result is the opposite of what the trivial
expectation is.

Not a biggie, this doesn't matter for my testing and could have been
fixed by just reprogramming the Daisy Seed too, but it's obviously a
"D'oh" moment.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>