Update plain 1590A board
I had decided that the effect board didn't need the input to be
ground-biased, and that was very wrong, and made some of my old effect
boards not work.  So add the AC couple capacitor and resistor to ground.

Also do the same boring RC power smoothing with the voodoo ferrite for
the +VA line.  Let's keep it simple and realistic - the noise isn't some
60Hz ripple, it's all SMPS high-frequency stuff.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>