Play silly buggers with the 1590LB enclosure
Yes, yes, I know I said I was thinking of a remote switch pedal in that
1590LB enclosure, but what a boring thing that would be.  So instead,
this fits an actual boost pedal into a 1590LB enclosure, using what is
basically a discrete component op-amp circuit.

It is not unlike my original boost pedal in the 1590A enclosure, except
the soft-clipping diodes are hardcoded - no room for a switch - and it's
a discrete JFET input buffer with a discrete op-amp-like thing instead
of using an actual op-amp.

Actual real op-amps are both simpler and much more powerful.  This just
approximates the real thing as a proof of concept.

It's going all out with a long-tailed pair, current sources, and a
push-pull output stage.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>