Alternate boost circuit layout
I'm not sure this fits vertically in a 1590LB, but the "everything on
top" model means that the audio jacks are very high up.  Which is not
great from a hole drilling standpoint.

So try putting the audio and DC power jacks on the back of the board,
moving the surface-mount components to the front.

And to make things fit - and because I feel like having done an op-amp
using discrete components - go back to just using an actual op-amp now.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>