Add spice model wrapper for opa1655 op-amp
The actual real spice file needs to be downloaded from TI, but this is
the "make that oddly formatted file JustWork(tm)" wrapper thing.

And yeah, ngspice has tons of issues with the TI opamp spice files, and
this all simulates much less reliably than the discrete component
version.  That's not a new issue.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>