Add a very early RP2354A / TAC5112 board design
I wanted to do some mixed-signal board, and decided that the RP2354A was
the thing to use, since it doesn't need to bother with external flash
etc.  You'd probably still want external PSRAM, but that's not a hard
requirement either.

The Raspberry Pi is well documented and this largely uses the design
from the "minimal board" reference docs.  That includes the KiCad symbol
itself for the RP2354A, which is in the example RPI schematic.

The codec is the TAC5112.  Not because I know what I'm doing, but
because it looks like a reasonably simple-to-use single-supply audio
codec that can use the same 3.3V supply that the RP2453 uses, keeping
things fairly simple.  Famous last words.

And no, none of this is really wired up yet, so there's no real PCB
layout yet.  I've not verified that any of the GPIOs make sense etc, and
all of that will presumably change when I do, and when I try to actually
lay things out.

I'll also need to actually do a proper job of picking footprints.  This
is mostly 0402 components with some larger caps randomly other sizes.
Very rough kind of "maybe something like this will fit and work"

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>