Update RP2354 board with proper inductor symbol
.. and fix up the LED simulation a bit too.  I simulate LEDs with the
built-in diode model, but with N ("emission coefficient") set to 3,
which gets me a forward voltage around 2V.

But I think I should change RS ("series resistance") instead (or in
addition to).  See

    https://led-ld.nichia.co.jp/api/data/spec/tech/SP-QR-C2-210932-2-E_The%20SPICE%20Model%20of%20LEDs.pdf

which talks much more about this.

I guess I don't care that much, I just wanted the current simulation for
the power LED to be in roughly the right ballpark.

The inductor change is the relatively important thing here.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>