First cut at BOM selection for caps and resistors
The size choices were a bit random, this standardizes on 0402 for
resistors.  Also, optimize the BOM by just sharing values where they
don't matter that much.

For example, we need those 5k1 resistors for the USB CC pulldowns, and
we might as well then use that for the i2c pullup resistors too.  And
the self-biasing source resistor for the JFET input could use the same.

For similar reasons, just standardize on 33R for the low value resistors
that we use for the USB data line serial termination and the analog
voltage RC filtering.

Capacitors are 0402 for small values (0.1uF and smaller) or 0603 for
bigger capacitance values.  That seems to match availability.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>