Switch the MMBF5103 JFET over to a OPA365 op amp
This is a long time coming, but it was triggered by jlcpcb not stocking
the MMBF5103, and all the alternatives looked worse.  So I guess sanity
has to prevail, and I should use an opamp rather than a discrete JFET
input stage.

I looked at the J201 - which I have tested and works, but has a
potentially very low V_{GSOFF} that will cause clipping.  And the
MMBFJ309L looked good in simulation and had better parameter values, but
I've never tested it, so it seemed a bit risky particularly since the
intended market seems to be for much higher frequency circuits.

I even looked at the JFE150, which is another silly-expensive thing sold
to audio markets.  I guess that tracks.

In the end, a plain simple opamp is definitely the sane choice.

To rebias the incoming signal, this uses the MICBIAS from the TAC5112,
which is typically 2.75V (but can be programmed to 2.5V or turned off
entirely etc).  It's there, it's almost what it is meant for, so let's
use it.

The OPA365 looks fine, and does rail-to-rail input and output and can do
single supplies down to 2.2V, but rather than put even more on the 3.3V
rail (and limiting the voltage swing), this just uses the 5V USB VBUS
power supply.

This does have some minimal caps for filtering (replacing the old +VA
supply to the JFET) but the OPA365 also claims good PSRR.

This all does seem like the right thing to do, but I'm still a bit sad
since the discrete JFET was just more interesting.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>